Adaptación a BBDD PostgreSQL

+++ scripts/bbdd/3.5/postgresql/update_from_3.4/sistra_redose_update_from_3.4.sql
@@ -1,4 +1,4 @@
-alter table RDS_DOCUM ADD DOC_DOCFOR  NUMBER(20);
+alter table RDS_DOCUM ADD DOC_DOCFOR  bigint;
 alter table RDS_DOCUM ADD DOC_ESFORM  VARCHAR2(1) default 'N' not null;
 
 alter table RDS_DOCUM add constraint RDS_DOCDOC_FK foreign key (DOC_DOCFOR) references RDS_DOCUM (DOC_CODIGO);
